Feat/front graph (#59)
* modify backend/main.py

* modify backend/models.py

* add backend/network_router.py

* modify frontend/app.css

* modify frontend/app.jsx

* modify frontend/index.css

* add nodes selection

* fix: graph page location

* Add graph optimization page and node selection

* add: working version

* fix: route display

---------

+++ project 3/backend/models/models.py
@@ -40,7 +40,7 @@ class TimeWindow(BaseModel):
 class Route(BaseModel):
     path: List[Edge]
     cost: float
-    time: int
+    time: float
     distance: float
